Fixed commands

+++ src/main/kotlin/com/blitzoffline/bountysystem/bounty/Bounty.kt
@@ -1,3 +1,3 @@
 package com.blitzoffline.bountysystem.bounty
 
-class Bounty(var id: Int, var payer: String, var target: String, var amount: Int, var placedTime: Long)
+class Bounty(var id: Short, var payer: String, var target: String, var amount: Int, var placedTime: Long)
